backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
pymysql>=1.0.0
aiofiles>=23.0.0
pytest>=7.0.0
pytest-asyncio>=0.20.0